Fixed Windows hosted mode bug. Moved the api from floats to doubles, and for fixed a "float/double coerced to integer string" bug.
Review by: jgw

+++ trunk/src/com/google/gwt/widgetideas/graphics/client/CanvasGradient.java
@@ -27,5 +27,5 @@ public abstract class CanvasGradient {
    * @param offset the offset at which to apply the colorstop
    * @param color the color of the color stop
    */
-  public abstract void addColorStop(float offset, Color color);
+  public abstract void addColorStop(double offset, Color color);
 }
